chore: step1과 step2 시작 파일 위치 변경
