+++ src/controller/web/startLottoGame.js
@@ -1,25 +1,32 @@
-import LottoMachine from "../../domain/LottoMachine/LottoMachine.js";
-import OutputView from "../../view/OutputView.js";
-import profitCalculator from "../../domain/profitCalculator/profitCalculator.js";
-import generateAnswerLotto from "../../domain/generateAnswerLotto.js";
-import InputHandler from "./InputHandler.js";
+import DomUpdator from "../../utils/DomUpdator";
+import DomSelector from "../../utils/domSelector";
+import handlePurchase from "./handlePurchase";
+import handleRestart from "./handleRestart";
+import handleWinningCheck from "./handleWinningCheck";
 
 const startLottoGame = async () => {
-  const purchaseAmount = await InputHandler.purchaseAmount();
-  const lottoPack = LottoMachine(purchaseAmount);
-  OutputView.purchaseCount(lottoPack.count);
-  OutputView.lottoPack(lottoPack.lottos);
+  const purchase_button = DomSelector.purchaseButton;
+  const purchase_amount = DomSelector.purchaseAmount;
+  const reuslt_button = DomSelector.reusltButton;
+  const lotto_result_modal = DomSelector.lottoResultModal;
+  const restart_button = DomSelector.restartButton;
 
-  const { winningNumbers, bonusNumber } = await InputHandler.answerLotto();
-  const answerLotto = generateAnswerLotto(winningNumbers, bonusNumber);
+  purchase_amount.addEventListener("keydown", (event) => {
+    if (event.key === "Enter") {
+      purchase_button.click();
+    }
+  });
 
-  const winningResult = lottoPack.compareAndReturnResult(answerLotto);
-  OutputView.winningStatistics(winningResult);
+  purchase_button.addEventListener("click", () => {
+    const { purchaseAmount, lottoPack } = handlePurchase();
+    reuslt_button.addEventListener("click", () => handleWinningCheck(purchaseAmount, lottoPack));
+  });
 
-  const profitRate = profitCalculator(purchaseAmount, winningResult);
-  OutputView.profitRate(profitRate);
+  lotto_result_modal.addEventListener("click", (event) => {
+    if (event.target === event.currentTarget) DomUpdator.showModal(lotto_result_modal, false);
+  });
 
-  await InputHandler.reStart(startLottoGame);
+  restart_button.addEventListener("click", () => handleRestart());
 };
 
 export default startLottoGame;
